renaming stratAbi to vaultAbi
